Removing "workshop"
Calling it an event, so that the word workshop is avoided, to help disambiguate, because Nodeschool calls their tutorials workshoppers. Also, used the word "do" instead of "run" because the word "run" is used to mean running an event in the preceding sentence. Too many changes of context. Run, run.